Update: PopUp Filter Code Refactor
